Bug 10607 - "play next track" doesn't work after I used the “songbird”search engine
Fixes the return values for the next and previous methods of sbPlaylistPlayback JS object. They were returning the dataremote object instead of the actual integer value, which caused XPConnect heartburn.

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@9866 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ components/playlistplayback/src/sbPlaylistPlayback.js
@@ -1155,7 +1155,7 @@ PlaylistPlayback.prototype = {
       throw Components.results.NS_ERROR_NOT_INITIALIZED;
     this._playNextPrev(1);
     this._controlTriggered.stringValue = Date.now();
-    return this._playlistIndex;
+    return this._playlistIndex.intValue;
   },
 
   previous: function() {
@@ -1164,7 +1164,7 @@ PlaylistPlayback.prototype = {
       throw Components.results.NS_ERROR_NOT_INITIALIZED;
     this._playNextPrev(-1);
     this._controlTriggered.stringValue = Date.now();
-    return this._playlistIndex;
+    return this._playlistIndex.intValue;
   },
 
   current: function() {
